requirements.txt
PyYAML==5.4.1
yamllint==1.26.3
pytest==6.2.5
google-cloud-bigquery==3.3.0
google-cloud-bigquery-storage==2.13.2
google-cloud-bigquery-datatransfer